Reorganize the rpc code
